% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor_curve.R
\name{survivor_curve}
\alias{survivor_curve}
\title{survivor_curve}
\usage{
survivor_curve(input_dir = here::here("raw data"), output_dir = "",
  group = list(sex = "M", ETH.group = "WHO", year = 2040),
  make_plot = TRUE)
}
\arguments{
\item{group}{}
}
\description{
survivor_curve
}
\examples{

}
